Move SplitButton to UIUtil::Widget
